+++ code/mc_1_21_4/src/main/java/org/betonquest/betonquest/mc_1_21_4/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * The BetonQuest Minecraft 1.21.4 Compatibility.
+ */
+@DefaultQualifier(value = NotNull.class, locations = {FIELD, PARAMETER, RETURN})
+package org.betonquest.betonquest.mc_1_21_4;
+
+import org.checkerframework.framework.qual.DefaultQualifier;
+import org.jetbrains.annotations.NotNull;
+
+import static org.checkerframework.framework.qual.TypeUseLocation.FIELD;
+import static org.checkerframework.framework.qual.TypeUseLocation.PARAMETER;
+import static org.checkerframework.framework.qual.TypeUseLocation.RETURN;